Extract Queryable into its own file
